+++ Finanzas/Entities/Transaction.cs
@@ -1,0 +1,10 @@
+﻿namespace Finanzas.Entities
+{
+    public class Transaction
+    {
+        public DateTime Date { get; set; }
+        public string Description { get; set; }
+        public decimal Amount { get; set; }
+        public string Type { get; set; }
+    }
+}
