+++ Finanzas/Entities/TransactionTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace Finanzas.Entities
+{
+    public enum TransactionTypes
+    {
+        Ingreso = 0,
+        Gasto = 1
+    }
+}
