+++ Finanzas/Infrastructure/ExtensionsMethods.cs
@@ -1,0 +1,10 @@
+﻿namespace Finanzas.Infrastructure
+{
+    public static class ExtensionsMethods
+    {
+        public static string ToEnumText<T>(this T tipoEnum) where T : struct
+        {
+            return tipoEnum.ToString();
+        }
+    }
+}
